Finish the append function
I've found the problem. It was in the condition of my while loop. I was
checking a step too far: looking for pointer !== null instead of
pointer.nextNode !== null. I needed my loop to stop at the element where
the next one was null, not the null itself. Damn how powerful it is to
come back after a break when you can't find the problem.